add new element (SiPM column no. in a layer) which is swSiPMID modulus number of sipms in a row

+++ lib/base/datastruct/SLookup.h
@@ -49,7 +49,7 @@ struct SIFI_EXPORT SLookupChannel
     virtual ~SLookupChannel() = default;
 
     ///@{
-    uint16_t m, l, s; ///< component of virtual address
+    uint8_t m, l, s; ///< component of virtual address
     ///}@
 
     virtual uint read(const char* buffer);
